acquisition/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "acquisition"
version = "0.1.0"
description = "Phase 9 of myavatar v6 — fulfil pending broll hints with Pexels API + text_card fallback"
requires-python = ">=3.11"
dependencies = [
    "pydantic>=2.0",
    "httpx>=0.27",
    "Pillow>=10.0",
    "python-dotenv>=1.0",
    "broll_resolver",
    "progress",
]

[project.scripts]
acquisition = "acquisition.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/acquisition"]
